fresh version
My first Hackathon - the idea was, using all the knowledge gained, to create a web page layout with certain functionality. In this case, the ability to navigate between pages that allows to demonstrate some actions of these pages  - a payment card widget(page) and sign in multi form for authorization on the home page.